Divide settings into multiple categories
The categories are: 'Graphics', 'Sounds', 'Controls' and 'Gameplay'.

Each category has its own scene containing all the options associated
with it. Only the 'Sounds' has been implemented so far, though.

The main contribution of this commit lies in the introduction of a new
prefab, titled 'BetterTabContainer'. As the name suggests, it behaves
similarly to the built-in tab container, i.e. only one of its children
can be visible at the same time and takes up all available space. That
being said, it doesn't actually use tabs, but a BoxContainer of buttons
which allows theming more appropriate to videogames. Unlike the built-in
implementation the new container also supports a vertical tabbar on the
left or right side (with the button text still being left-to-right).

This is a rather complex tool script and might still contains bugs!