Turn PauseMenu from autoload into a local instance
This makes more sense, as the PauseMenu does not have to be available at
all times. Its use is limited to the Game scene anyway.

Also, moves the code responsible for hiding the PauseMenu when switching
to another scene (e.g. the main menu) from the SceneSwitcher autoload to
the PauseMenu itself. There's no point in checking for a relatively rare
event (changing away from the pause menu) on *every* scene change.
The menu has to be hidden *before* changing the scene (or the connected
`_on_visibility_changed` callback won't be able to access the tree, and
reset its `paused` variable back to false).

Next, moves the `_unhandled_input` method that brings up the PauseMenu
when the "pause_game" action is pressed, to the PauseMenu as well. That
way the scene can be simply instanced anywhere it's needed and will work
right away, without the need to add any custom code to trigger it.
This requires its `process_mode` to be set to `PROCESS_MODE_ALWAYS`.

Lastly, renames the "Resume" button to "ResumeGame", for added clarity.
However, the button's text remains unchanged, as the the header's text
("Game Paused") should give sufficient context here.