Add default audio bus layout & volume settings
Adds a default bus layout, consisting of four different buses:
- Master
- Music
- SFX
- UI

The settings scene will automatically add a volume setting for each of
those buses, which show the bus' volume in percent, and include buttons
to increase and decrease the volume as well as a toggle to mute the bus.
If one adds or remove a bus from the bus layout, the settings scene will
reflect that, without any further user-side changes needed.

Note: The settings are *not* (yet) saved to disk.